wrote the paper (text + img)
it did not compile the pdf book for me although html worked. potentially I made some errors?
potentially the equation in line 46-54, uninterpretable characters in the bib file or hyperlinks create problems.
not entirely sure. I can change stuff or figure it out if it turns out to be indeed a problem.